Issue #252 extract out common offheap and disk store tests
